// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ClickRow: {
            /** Format: date-time */
            ts: string;
            code: string;
            prefix?: string;
            access_prefix?: string;
            long_url?: string;
            owner: string;
            campaign_code?: string;
            ctv_ref?: string;
            ctv_display?: string;
            device: string;
            os: string;
            browser: string;
            source_group: string;
            referer_host?: string;
            country: string;
            province?: string;
            /** @description Che với vai trò khác admin */
            ip: string;
            is_bot: boolean;
            is_suspicious: boolean;
            is_repeat: boolean;
        };
// ...
    };
// ...
}